+++ packages/spacecat-shared-dynamo/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-dynamo-v1.3.24](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.23...@adobe/spacecat-shared-dynamo-v1.3.24) (2024-06-15)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#266](https://github.com/adobe/spacecat-shared/issues/266)) ([a89ab83](https://github.com/adobe/spacecat-shared/commit/a89ab83e1c108c10044f6d098526bafed67b88b2))
+
 # [@adobe/spacecat-shared-dynamo-v1.3.23](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.22...@adobe/spacecat-shared-dynamo-v1.3.23) (2024-06-08)